Tier B Phase 3d-ii-a: relocate detection/parsing utils to the package
First step of the renderer/cluster relocation. The pure UI-type detection +
tool-result parsing utils move into @mcpjam/widget-react with inspector shims.

- tool-result-utils (readToolResult{Object,Meta,ServerId}) -> package (pure).
- mcp-apps-utils core (UIType, detectUIType, detectUiTypeFromTool,
  getUIResourceUri + tool-visibility re-exports) -> package.
- Inspector @/lib/tool-result-utils + @/lib/mcp-ui/mcp-apps-utils become
  re-export shims; the ListToolsResultWithMetadata-typed isMCPApp/isOpenAIApp/
  isOpenAIAppAndMCPApp helpers stay inspector-side.
- Package gains @mcp-ui/client + @modelcontextprotocol/client deps, externalized
  in tsup (dts has external refs only).

Verified: package guard/typecheck/test/build; inspector typecheck:client + 232
tests; root typecheck. No behavior change.
