+++ widget-react/package.json
@@ -35,7 +35,9 @@
     "react-dom": "^19.0.0"
   },
   "dependencies": {
+    "@mcp-ui/client": "^5.9.0",
     "@mcpjam/sdk": "^1.17.0",
+    "@modelcontextprotocol/client": "^2.0.0-alpha.2",
     "@modelcontextprotocol/ext-apps": "^1.7.2"
   },
   "devDependencies": {
